Refactor RocketChat user operations
